Created fragments and data packages
