--- app/src/main/java/com/example/calhamnorthway/group17projectpart4/Gender.java
@@ -1,3 +1,0 @@
-package com.example.calhamnorthway.group17projectpart4;
-
-public enum Gender{ Male, Female };
