+++ app/src/main/java/com/example/calhamnorthway/group17projectpart4/data/Gender.java
@@ -1,0 +1,3 @@
+package com.example.calhamnorthway.group17projectpart4.data;
+
+public enum Gender{ Male, Female };
