+++ app/src/main/java/com/example/calhamnorthway/group17projectpart4/fragments/MeetPeopleFragment.java
@@ -1,12 +1,13 @@
-package com.example.calhamnorthway.group17projectpart4;
+package com.example.calhamnorthway.group17projectpart4.fragments;
 
 import android.content.Context;
-import android.net.Uri;
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import com.example.calhamnorthway.group17projectpart4.R;
 
 
 /**
